+++ parler/managers.py
@@ -33,7 +33,7 @@ class TranslatableQuerySet(QuerySet):
         Set the language code to assign to objects retrieved using this QuerySet.
         """
         if language_code is None:
-            language_code = appsettings.PARLER_DEFAULT_LANGUAGE_CODE
+            language_code = appsettings.PARLER_LANGUAGES.get_default_language()
 
         self._language = language_code
         return self