+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes in version 1.0 (in development)
   applications which use ``translations = TranslatedFields(..)`` in their models.
 * Fix supporting different database (using=) arguments.
 * Fix list language, always show translated values.
+* Fix ``is_supported_django_language()`` to support dashes too
 
 
 Changes in version 0.9.4 (beta)