+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes in version 1.0 (in development)
   This only affects your South migrations. Use ``manage.py schemamigration appname --empty "upgrade_to_django_parler10"`` to upgrade
   applications which use ``translations = TranslatedFields(..)`` in their models.
 * Allow ``.translated()`` and ``.active_translations()`` to filter on translated fields too.
+* Added ``language_code`` parameter to ``safe_translation_getter()``, to fetch a single field in a different language.
+* Added partial support for tabs on inlines when the parent object isn't a translated model.
 * Fix supporting different database (using=) arguments.
 * Fix list language, always show translated values.
 * Fix ``is_supported_django_language()`` to support dashes too