+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes in version 1.0 (in development)
 * Added a ``ViewUrlMixin`` so views can tell the template what their exact canonical URL should be.
 * Added ``TranslatableCreateView`` and ``TranslatableUpdateView`` views, and associated mixins.
 * Fix missing "language" GET parmeter for Django 1.6 when filtering in the admin (due to the ``_changelist_filters`` parameter).
+* Support missing `SITE_ID` setting for Django 1.6.
 
 
 Released on 1.0a1: